requirements.txt
pathway
nats-py